+++ main.h
@@ -31,7 +31,7 @@ int _printf_oct(va_list *list, char *ptr, int *k);
 void _printf_oct_rec(int *n, unsigned int num, char *ptr, int *k);
 
 int _printf_hex(va_list *list, char *ptr, int *k);
-void _printf_hex_rec(int *n, unsigned int num, char *ptr, int *k);
+void _printf_hex_rec(int *n, unsigned long num, char *ptr, int *k);
 
 int _printf_HEX(va_list *list, char *ptr, int *k);
 void _printf_HEX_rec(int *n, unsigned int num, char *ptr, int *k);
@@ -39,6 +39,7 @@ void _printf_HEX_rec(int *n, unsigned int num, char *ptr, int *k);
 void Buffer_editor(char *ptr, int *k, char c);
 
 int printf_custom_S(va_list *list, char *p, int *k);
+
 
 
 #endif /* MAIN_H */
